+++ src/components/TableHeadCell.js
@@ -43,7 +43,7 @@ const defaultHeadCellStyles = {
   },
 };
 
-class MUIDataTableHeadCell extends React.Component {
+class TableHeadCell extends React.Component {
   static propTypes = {
     /** Extend the style applied to components */
     classes: PropTypes.object,
@@ -112,4 +112,4 @@ class MUIDataTableHeadCell extends React.Component {
   }
 }
 
-export default withStyles(defaultHeadCellStyles, { name: "MUIDataTableHeadCell" })(MUIDataTableHeadCell);
+export default withStyles(defaultHeadCellStyles, { name: "MUIDataTableHeadCell" })(TableHeadCell);
