+++ src/components/TableHeadRow.js
@@ -8,7 +8,7 @@ const defaultHeadRowStyles = {
   root: {},
 };
 
-class MUIDataTableHeadRow extends React.Component {
+class TableHeadRow extends React.Component {
   static propTypes = {
     /** Extend the style applied to components */
     classes: PropTypes.object,
@@ -28,4 +28,4 @@ class MUIDataTableHeadRow extends React.Component {
   }
 }
 
-export default withStyles(defaultHeadRowStyles, { name: "MUIDataTableHeadRow" })(MUIDataTableHeadRow);
+export default withStyles(defaultHeadRowStyles, { name: "MUIDataTableHeadRow" })(TableHeadRow);
