+++ src/components/TableSelectCell.js
@@ -27,7 +27,7 @@ const defaultSelectCellStyles = {
   disabled: {},
 };
 
-class MUIDataTableSelectCell extends React.Component {
+class TableSelectCell extends React.Component {
   static propTypes = {
     /** Select cell checked on/off */
     checked: PropTypes.bool.isRequired,
@@ -62,4 +62,4 @@ class MUIDataTableSelectCell extends React.Component {
   }
 }
 
-export default withStyles(defaultSelectCellStyles, { name: "MUIDataTableSelectCell" })(MUIDataTableSelectCell);
+export default withStyles(defaultSelectCellStyles, { name: "MUIDataTableSelectCell" })(TableSelectCell);
